+++ crates/nostr-relay-pool/src/relay/mod.rs
@@ -405,14 +405,14 @@ impl Relay {
 
     /// Send msg to relay
     #[inline]
-    pub fn send_msg(&self, msg: ClientMessage<'_>) -> Result<(), Error> {
-        self.inner.send_msg(msg)
+    pub async fn send_msg(&self, msg: ClientMessage<'_>) -> Result<(), Error> {
+        self.inner.send_msg(msg).await
     }
 
     /// Send multiple [`ClientMessage`] at once
     #[inline]
-    pub fn batch_msg(&self, msgs: Vec<ClientMessage<'_>>) -> Result<(), Error> {
-        self.inner.batch_msg(msgs)
+    pub async fn batch_msg(&self, msgs: Vec<ClientMessage<'_>>) -> Result<(), Error> {
+        self.inner.batch_msg(msgs).await
     }
 
     async fn _send_event(
@@ -422,7 +422,8 @@ impl Relay {
     ) -> Result<(bool, String), Error> {
         // Send the EVENT message
         self.inner
-            .send_msg(ClientMessage::Event(Cow::Borrowed(event)))?;
+            .send_msg(ClientMessage::Event(Cow::Borrowed(event)))
+            .await?;
 
         // Wait for OK
         self.inner
@@ -572,7 +573,7 @@ impl Relay {
             .await;
 
         // Send REQ message
-        if let Err(e) = self.inner.send_msg(msg) {
+        if let Err(e) = self.inner.send_msg(msg).await {
             // Remove previously added subscription
             self.inner.remove_subscription(&id).await;
 
@@ -600,7 +601,7 @@ impl Relay {
         };
 
         // Send REQ message
-        self.inner.send_msg(msg)?;
+        self.inner.send_msg(msg).await?;
 
         // No auto-close subscription: update subscription filter
         self.inner.update_subscription(id, filters, true).await;
@@ -688,7 +689,7 @@ impl Relay {
             subscription_id: Cow::Borrowed(&id),
             filter: Cow::Owned(filter),
         };
-        self.inner.send_msg(msg)?;
+        self.inner.send_msg(msg).await?;
 
         let mut count = 0;
 
@@ -714,7 +715,7 @@ impl Relay {
         .ok_or(Error::Timeout)?;
 
         // Unsubscribe
-        self.inner.send_msg(ClientMessage::close(id))?;
+        self.inner.send_msg(ClientMessage::close(id)).await?;
 
         Ok(count)
     }